refactor: move src/http/handlers to src/adapters/http/handlers
