+++ src/adapters/http/handlers/executionResult.stub.ts
@@ -1,13 +1,13 @@
 import type { FastifyReply, FastifyRequest } from "fastify";
-import { SCHEMA_VERSION, type ExecutionResultResponse } from "../../contract/v1/types.js";
-import { type LedgerStore } from "../../ledger/store.js";
+import { SCHEMA_VERSION, type ExecutionResultResponse } from "../../../contract/v1/types.js";
+import { type LedgerStore } from "../../../ledger/store.js";
 import {
   LEDGER_ERROR_CODES,
   LedgerWriteError,
   writeExecutionResultLedgerEntry
-} from "../../ledger/writer.js";
-import { parseExecutionResultRequest } from "../../contract/v1/validation.js";
-import { ContractValidationError } from "../errors.js";
+} from "../../../ledger/writer.js";
+import { parseExecutionResultRequest } from "../../../contract/v1/validation.js";
+import { ContractValidationError } from "../../../contract/v1/errors.js";
 
 export const createExecutionResultStubHandler = (store: LedgerStore) => {
   return async (request: FastifyRequest, reply: FastifyReply) => {
