+++ src/adapters/http/handlers/srLevelsV2Ingest.ts
@@ -3,7 +3,7 @@ import {
   parseSrLevelsV2IngestRequest,
   type SrLevelsV2IngestCreatedResponse,
   type SrLevelsV2IngestAlreadyIngestedResponse
-} from "../../contract/v2/srLevels.js";
+} from "../../../contract/v2/srLevels.js";
 import {
   V2ContractValidationError,
   V2_SCHEMA_VERSION,
@@ -12,9 +12,9 @@ import {
   unauthorizedV2Error,
   buildSrThesisV2ConflictEnvelope,
   internalErrorV2
-} from "../../contract/v2/errors.js";
-import { SrThesesV2Store, SrThesisV2ConflictError } from "../../ledger/srThesesV2Store.js";
-import { safeEqual } from "../../adapters/http/auth.js";
+} from "../../../contract/v2/errors.js";
+import { SrThesesV2Store, SrThesisV2ConflictError } from "../../../ledger/srThesesV2Store.js";
+import { safeEqual } from "../auth.js";
 
 const ENV_VAR = "OPENCLAW_INGEST_TOKEN";
 const HEADER = "x-ingest-token";
