+++ src/http/routes.ts
@@ -1,18 +1,18 @@
 import type { FastifyInstance, FastifyReply } from "fastify";
 import { buildOpenApiDocument } from "../adapters/http/openapi.js";
-import { createClmmExecutionResultHandler } from "./handlers/clmmExecutionResult.js";
-import { createExecutionResultHandler } from "./handlers/executionResult.js";
-import { createPlanHandler } from "./handlers/plan.js";
-import { createWeeklyReportHandler } from "./handlers/report.js";
-import { createCandlesIngestHandler } from "./handlers/candlesIngest.js";
-import { createRegimeCurrentHandler } from "./handlers/regimeCurrent.js";
-import { createSrLevelsIngestHandler } from "./handlers/srLevelsIngest.js";
-import { createSrLevelsCurrentHandler } from "./handlers/srLevelsCurrent.js";
-import { createInsightsIngestHandler } from "./handlers/insightsIngest.js";
-import { createInsightsCurrentHandler } from "./handlers/insightsCurrent.js";
-import { createInsightsHistoryHandler } from "./handlers/insightsHistory.js";
-import { createSrLevelsV2IngestHandler } from "./handlers/srLevelsV2Ingest.js";
-import { createSrLevelsV2CurrentHandler } from "./handlers/srLevelsV2Current.js";
+import { createClmmExecutionResultHandler } from "../adapters/http/handlers/clmmExecutionResult.js";
+import { createExecutionResultHandler } from "../adapters/http/handlers/executionResult.js";
+import { createPlanHandler } from "../adapters/http/handlers/plan.js";
+import { createWeeklyReportHandler } from "../adapters/http/handlers/report.js";
+import { createCandlesIngestHandler } from "../adapters/http/handlers/candlesIngest.js";
+import { createRegimeCurrentHandler } from "../adapters/http/handlers/regimeCurrent.js";
+import { createSrLevelsIngestHandler } from "../adapters/http/handlers/srLevelsIngest.js";
+import { createSrLevelsCurrentHandler } from "../adapters/http/handlers/srLevelsCurrent.js";
+import { createInsightsIngestHandler } from "../adapters/http/handlers/insightsIngest.js";
+import { createInsightsCurrentHandler } from "../adapters/http/handlers/insightsCurrent.js";
+import { createInsightsHistoryHandler } from "../adapters/http/handlers/insightsHistory.js";
+import { createSrLevelsV2IngestHandler } from "../adapters/http/handlers/srLevelsV2Ingest.js";
+import { createSrLevelsV2CurrentHandler } from "../adapters/http/handlers/srLevelsV2Current.js";
 import type { ClockPort } from "../application/ports/clock.js";
 import type { GetCurrentRegimeUseCase } from "../application/use-cases/getCurrentRegimeUseCase.js";
 import type { GeneratePlanUseCase } from "../application/use-cases/generatePlanUseCase.js";
